Parse pre Core34 rules correct and enable change logging.
This fixes bug #0000546 and older rules should be parseable by webgui
the values are filled up with default when missing. As user requested
the ability to enable logging for mode1 rules was added, earlier we
were defaulting to logging false. It is recommended to check all
outgoing rule if working correct or just reset rules and generate a
new set of rules based on core34 features.