Fixed OutgoingFW Parser, all option removed
To be compatible with old rules all will be translated to tcp and udp
like in earlier versions of outgoingfw. So all is not doing what you
expect, but if we change this all old rules using all and ports need to
be translated to udp or tcp.
When using GRE and ESP ports will be ignored.