Fixed the urlfilter bug #0000552
The config recently contained all filter destinations thats why the squidGuard
exhausts the memory for each filter process even when the filter destination was
not used by any rule.

The destinations were added because time constraints were not able to add needed
destinations, so all destination were added by the config parser.

Now the config parser checks if the destination is enabled by generel, if not it
checks if the destination is needed by a time constraint.