installer: add parameter to skip a block device for installation.
A hyper-v user has reported that the ipfire installer detects a wrong drive
as sda and always try to install on this.
Now you can boot the installation with "ipfire skipdst=sda" to install
this may help to install on hyper-v and other situations.
If more than one should skipped add a parameter for every drive.