index.cgi: Show OpenVPN header if server is disabled but N2N connections are used.
If the ovpn server for roadwarrior connections was stopped, there was no OpenVPN header
displayed on the index.cgi. If there are configured Net-to-Net connections, they
could be established in the background but their  status can be found
on the VPN (IPsec) section which is wrong. If there is also no IPsec
configured there aren't any information about the Net-to-Net connections on the start page.