squid: Speed up the stop process.
The squid proxy server has been blocked by unresponsive redirectors and it took ages to kill it in the past.

To speed up the shutdown process, we will stop all redirector services at the same time. If the squid service
is still running we will wait up to 30 seconds before a TERM and finaly a KILL signal will be sent.

Fixes #10368.