multicat: add a new addon package "multicat" - The multicast swiss knife
Very useful for analyzing multicast traffic directly on the router/
firewall without the need for a large software like vlc or the like.

http://www.videolan.org/projects/multicat.html

Simple and efficient multicast and transport stream manipulation

The multicat package contains a set of tools designed to easily and
efficiently manipulate multicast streams in general, and MPEG-2
Transport Streams (ISO/IEC 13818-1) in particular.

The multicat suite of applications is very lightweight and designed
to operate in tight environments. Memory and CPU usages are kept to
a minimum, and they feature only one thread of execution.

multicat needs bitstream as a build dependency
http://www.videolan.org/developers/bitstream.html