Increase performance of the squidclamav redirector.
To boost up the performance, now we trust the proxy cache.

I add some changes to the proxy.cgi to configure the proxy and the
squidclamav in the right way.

I also add a hook that allows us to generate a new configuration
if the cgi script will be launched from the shell.

Fixes #10367.