Core 76: Don't extract firewall config files.
While updating the configfiles of the firewall has been extracted and replaced,
the existing one's if the new firewall was allready installed. As a result of
this behaviour all firewall rules, settings or created groups were lost. We now
try to create them by using touch if they do not exist.