updxlrator: Handle colon in filenames in updatexlrator.cgi
The updatexlrator.cgi is assuming that colon is not part of any
filenames downloaded by updxlrator. Fix this by fixing regular
expression used for listing pending download files, and by
fixing regular expression used to listing files in maintenance view.

The fix for listing in maintenance is not very nice, I just use
":separator:" as field separator instead of just ":", which was used
earlier. This should make it very unlikely to collide with a part of
the filename.

Also make the files in download be sorted using lower case, like it is
done for files in the vendor path.

The problem can be reproduced by downloading
http://mirror.us.leaseweb.net/archlinux/community/os/x86_64/
docker-1:0.7.6-1-x86_64.pkg.tar.xz, after pkg.tar.xz files support has
been added to updxlrator.