firewalllog.dat: Add First and Last links, for fast navigation in log lines
Add two links for navigating in log pages, one for going to the
first page and one for goint to the last page.

Introduce variable which points to the first log line to
display on the last possible page.

The commit contains hardcoded text for Last and First, and I
am also uncertain how spaces between links should be handled.
So this should be improved based on feedback.