showrequestfromcountry.dat: Use language string, and fix links
Define language key for input field.
Fix links for older and newer links.
Indentation fixes.

The code is a copy from showrequestfromip.dat, ideally
we should have merged all three showrequestfrom*.dat files
into one file, but I do not do that now, because it would
really require a rewrite of most of the logic, and I understand
that one does not want to do such changes in 2.x.