webgui menus: Make top menu item clickable on mobile safari
The menus in the new default theme does not work properly on
ipad and I guess other devices using mobile safari.

The reason is that <a> tags without href is not marked as
clickable by ios, and therefore no css hover effect is applied.

The fix is either to add a href="#" attribute
or a onclick="return true;" atribute on the <a> tag.
I prefer the href solution, since it is a no javascript solution.