index.cgi: Make html valid, and improve dialup and vpn display
Make the html validate. One part of the changes is to
move style from using deprecated attributes to using
style attribute on tag.
The other part is to make sure that tables, rows and cells
are properly closed and nested.

Use a table for showing output from the dialctrl script.