qos and process graphs: avoid random colors
For the graphs for qos classes, and for process
count and memory usage, the colors change on
each page reload, since the colors are currently
totally random.

This change uses defined color indexes, 15 different
ones, that should be enough for qos at least, and is
enough for processes on my setup. This makes
the colors consistent between page loads, so it makes
it easier to monitor changes.

Perhaps one should define 30 different colors meant
to be used for graphing, then one could avoid random
colors also for the sensor graphs, and make it less
likely to get color collision for the process graphs
as well. Defining 30 such colors is not part of this
change.