rngd: Silence initscript when rngd is already started.
When a hardware random number generator is found by udev
it will start rngd automatically which is what we also do
by default in the initialisation sequence of the system
(e.g. for RDRAND).
The user will then see an error message that rngd has
already been started which was confusing.