ddns.cgi: Fix enable/disable handling of entries.
When the "enabled" checkbox is checked a "on" will be returned,
if the box is unchecked checkboxes will return nothing.

As a result of this behaviour the ddns.conf contained entries which have been disabled in the WUI.

We now check if the checkbox returns a "on", otherwise we will set the "enabled" value to "off" to
prevent from this problem.