Ship bash with Core Update 86 again
It seems that some systems installed some pre-versions
of the distribution which came with an outdated version
of bash. Since this update uses some scripts that use
bash-4-isms, we ship the shell once again.

An other reason is that there have been security fixes
in the shell which should be fixed on all systems.