nfs-server: Fix kernel version check in initscript.
nfsd requires a mounted nfsd filesystem which has been introduced in
the kernel 2.6 tree. To determine the current running kernel, a check
was included in the initscript which works fine until we switched to a kernel
version 3.x.

This commit fixes this check, so the nfs-server will startup again.

Fixes #10760.