setup: Has been moved to /usr/sbin
The setup program has been moved from /usr/local/sbin/setup
to /usr/sbin/setup to comply better with FHS. The old copy
was left in the file system on updated systems and was
preferred when called from the shell which is not intended.

The old file is now removed and to make sure every system
has got the new file it is shipped once again.