backup: Include sshd config file.
The sshd config file was not included in the backups, so if the port
was switched from the default one to another one, the changed value was
correctly restored and displayed in the WUI but still the original sshd
config file was present and the default port has been used.

This issue is fixed by this commit.

Fixes #10446.