dhcp: Enter edit mode after adding a static lease from list
After a lease has been added from the list of leases it is often
desirable to edit it right away. It appeared that the system
was in edit mode after the lease has been added because the form
had the values of the recently added lease, but hitting the
"Add" button cause an error message that stated that this is
a duplicate entry.

This patch will switch to edit more where the user can change
the values and hit "Update". If he or she chooses to ignore
that the entry is still saved.