pppsetup.cgi: Fix site layout when no TYPE is specified
There was an issue with false generated HTML code, in case
of an empty or unset $pppsettings{'TYPE'} variable which results
in a missplaced website footer.

This patch moves the code for closeing the table and the call of the
closebox() function to the correct place to prevent this display issue.

Fixes #10565.

Signed-off-by: Stefan Schantl <stefan.schantl@ipfire.org>
Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>