Add new package libusbredir
This package adds support for the use redirection of spice.
It is now possible to attach USB devices of the host where the spice
client run to the virtual machine.

The binary is not needed for this functionality and that's why they is
not shipped with the package

This feature is also enabled in qemu.

Signed-off-by: Jonatan Schlag <jonatan.schlag@ipfire.org>
Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>