Libvirt: Fix update.sh script
The virtlogd could only be restarted when the daemons run. The update.sh
script tried to restart the daemon no matter if the daemons run or not.
This behaviour produce problems.

An If statement now checks if the daemon runs or not and execute the
command that is suitable for the situation.

Fixes: #11172

Signed-off-by: Jonatan Schlag <jonatan.schlag@ipfire.org>
Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>