shadow-utils: Create standard set of configuration files
Previously we copied the default configuration from the upstream
package and modified that. Unfortunately a patch and a sed command
changed the file which resulted in unwanted changes.

This patch removes the patch and sed command and adds a new set
of configuration files that just need to be copied to the system.

Fixes #11195

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>